apiv2: cleanup types, add terms and service to register
